requirements.txt
pandas
selenium
webdriver-manager
python-dotenv